requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
apscheduler==3.10.4
pydantic-settings==2.5.2
pyyaml==6.0.2
